restructure: monorepo with code/dart, code/ts, code/py scaffolds
- Move Dart SDK into code/dart/, update pubspec.yaml repository URL
- Move doc/ to docs/ at root
- Scaffold TypeScript project (code/ts/) — @macss/service-client
- Scaffold Python project (code/py/) — macss-service-client
- Create cross-language test infrastructure (code/tests/)
- Add root .gitignore, LICENSE, README.md as monorepo overview

Closes #3
