+++ .gitignore
@@ -1,11 +1,2 @@
-# https://dart.dev/guides/libraries/private-files
-# Created by `dart pub`
-.dart_tool/
-
-# Avoid committing pubspec.lock for library packages; see
-# https://dart.dev/guides/libraries/private-files#pubspeclock.
-pubspec.lock
-
-# .dev 
-.dev/*
-!.dev/runbooks
+# developer-specific files
+.dev/
