+++ code/dart/example/example.dart
@@ -1,4 +1,4 @@
-import 'package:service_client/service_client.dart';
+import '../lib/service_client.dart';
 
 import 'controllers/todo_controller.dart';
