+++ code/dart/lib/service_client.dart
