+++ code/dart/lib/src/core/result.dart
