+++ code/dart/lib/src/core/service_core.dart
