+++ code/dart/lib/src/core/service_failure.dart
