+++ code/dart/lib/src/http/http_client.dart
