+++ code/dart/lib/src/http/http_exceptions.dart
