+++ code/dart/lib/src/http/http_service_client.dart
