+++ code/dart/test/core/result_test.dart
@@ -1,4 +1,4 @@
-import 'package:service_client/service_client.dart';
+import '../../lib/service_client.dart';
 import 'package:test/test.dart';
 
 void main() {
