+++ code/py/CHANGELOG.md
@@ -1,0 +1,10 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format loosely follows [Keep a Changelog](https://keepachangelog.com/)
+and the project adheres to [Semantic Versioning](https://semver.org/).
+
+## [Unreleased]
+
+- Project scaffolded — no source code yet.
