+++ code/py/README.md
@@ -1,0 +1,27 @@
+# service-client
+
+Service client abstraction with the Result pattern for Python.  
+Transport-agnostic interface with an HTTP implementation.
+
+> Also available in **Dart**: [service_client](https://pub.dev/packages/service_client) · **TypeScript**: [@macss/service-client](https://www.npmjs.com/package/@macss/service-client)
+
+---
+
+## Status
+
+This package is **not yet implemented**. The project has been scaffolded with configuration files only.  
+See the [Roadmap](../../docs/roadmap.md) for the planned timeline.
+
+---
+
+## Installation
+
+```bash
+pip install macss-service-client
+```
+
+---
+
+## License
+
+MIT © [ccisne.dev](https://ccisne.dev)
