+++ code/py/pyproject.toml
@@ -1,0 +1,45 @@
+[build-system]
+requires = ["setuptools>=68", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "macss-service-client"
+version = "0.0.1"
+description = "Service client abstraction with the Result pattern for Python. Transport-agnostic interface with an HTTP implementation."
+readme = "README.md"
+license = "MIT"
+requires-python = ">=3.11"
+authors = [{ name = "ccisne.dev" }]
+keywords = ["service-client", "result-pattern", "http", "macss"]
+classifiers = [
+    "Development Status :: 1 - Planning",
+    "Intended Audience :: Developers",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Typing :: Typed",
+]
+dependencies = []
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=8.0",
+    "httpx>=0.27",
+]
+
+[project.urls]
+Homepage = "https://github.com/macss-dev/service_client"
+Repository = "https://github.com/macss-dev/service_client/tree/main/code/py"
+Issues = "https://github.com/macss-dev/service_client/issues"
+Documentation = "https://github.com/macss-dev/service_client/tree/main/code/py#readme"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+
+[tool.pyright]
+pythonVersion = "3.11"
+typeCheckingMode = "strict"
