+++ code/py/src/service_client/__init__.py
@@ -1,0 +1,1 @@
+"""service_client — Service client abstraction with the Result pattern for Python."""
