+++ code/tests/integration_test/parity_test.ps1
@@ -1,0 +1,13 @@
+# Cross-language parity test runner
+# Validates that Dart, TypeScript, and Python implementations produce
+# identical results for shared test fixtures in ../fixtures/.
+#
+# Usage: pwsh code/tests/integration_test/parity_test.ps1
+#
+# Prerequisites:
+#   - Dart SDK, Node.js, Python 3.11+ installed
+#   - Each SDK built and dependencies resolved
+#
+# Status: placeholder — no fixtures or implementations yet.
+
+Write-Host "parity_test: no fixtures to test yet." -ForegroundColor Yellow
