+++ code/ts/.prettierignore
@@ -1,0 +1,3 @@
+dist/
+node_modules/
+package-lock.json
