+++ code/ts/README.md
@@ -1,0 +1,27 @@
+# service-client
+
+Service client abstraction with the Result pattern for TypeScript.  
+Transport-agnostic interface with an HTTP implementation.
+
+> Also available in **Dart**: [service_client](https://pub.dev/packages/service_client) · **Python**: [macss-service-client](https://pypi.org/project/macss-service-client/)
+
+---
+
+## Status
+
+This package is **not yet implemented**. The project has been scaffolded with configuration files only.  
+See the [Roadmap](../../docs/roadmap.md) for the planned timeline.
+
+---
+
+## Installation
+
+```bash
+npm install @macss/service-client
+```
+
+---
+
+## License
+
+MIT © [ccisne.dev](https://ccisne.dev)
